desafio07/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
botcity-crawler-plugin==0.2.2
botcity-framework-base==0.4.4
botcity-framework-web==0.9.0
botcity-maestro-sdk==0.5.1
botcity-utils==0.1.2
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
distro==1.9.0
fake-useragent==1.5.1
idna==3.8
importlib_metadata==8.5.0
lxml==5.3.0
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parse==1.20.2
pillow==10.4.0
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
selenium==3.141.0
six==1.16.0
soupsieve==2.6
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
w3lib==2.2.1
webdriver-manager==4.0.2
websockets==10.4
zipp==3.20.2
